Terminando associação das certidões

+++ cartorio/src/main/java/com/lmmartins/cartorio/services/OfficeService.java
@@ -1,50 +1,56 @@
 package com.lmmartins.cartorio.services;
 
+import com.lmmartins.cartorio.dto.Response;
 import com.lmmartins.cartorio.models.Office;
 import com.lmmartins.cartorio.repositories.IOfficeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class OfficeService {
     @Autowired
     private IOfficeRepository repository;
 
-    public void deleteOfficeById(Long officeId){
-//        Optional<Office> office = repository.findById(officeId);
-//        if(office.isEmpty()) {
-//            return notFound().build();
-//        }
-//        else {
+    public Response deleteOfficeById(Long officeId) {
+        Optional<Office> office = repository.findById(officeId);
+        if (office.isEmpty()) {
+            return new Response(false,404,"Not found the office.");
+        }
+        else {
             try {
                 repository.deleteById(officeId);
+                return new Response(true,200, null);
             } catch (Exception e) {
                 System.out.println(e.getMessage());
+                return new Response(false,500, "Internal error.");
             }
-//        }
-    }
-
-    public void updateOfficeById(Long officeId, Office newOffice){
-        Office updatedOffice = new Office(
-                officeId,
-                newOffice.getName(),
-                newOffice.getAddress()
-        );
-
-        try {
-            repository.save(updatedOffice);
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
         }
     }
 
-    public Long createOffice(Office office){
+    public Response updateOfficeById(Long officeId, Office office){
+        Office newOffice = new Office(
+                officeId,
+                office.getName(),
+                office.getAddress()
+        );
+
         try {
-            return repository.save(office).getId();
+            repository.save(newOffice);
+            return new Response(true, 200, null);
         } catch (Exception e) {
-            return null;
+            System.out.println(e.getMessage());
+            return new Response(false, 500, null);
+        }
+    }
+
+    public Response createOffice(Office office){
+        try {
+            return new Response(true, 200,null, repository.save(office).getId());
+        } catch (Exception e) {
+            return new Response(false, 500, null);
         }
     }
 
@@ -59,6 +65,14 @@ public class OfficeService {
     public Office getOfficeById(Long officeId){
         try {
             return repository.findById(officeId).get();
+        } catch (Exception e) {
+            return null;
+        }
+    }
+
+    public Boolean existOffice(Long officeId){
+        try {
+            return repository.existsById(officeId);
         } catch (Exception e) {
             return null;
         }
